backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-multipart
pillow
python-dotenv
alembic
python-jose[cryptography]
